requirements.txt
langchain>=1.2.2
langchain-google-genai>=4.1.3
langchain-chroma>=1.1.0
langchain-community>=0.3.21
pypdf>=6.5.0
gradio>=5.25.1
python-dotenv>=1.1.0
